Fixed predicate refactor at #5161;  * Interdict, Ouphe Vandals - fixed that it catch triggered + activated, but must only activated;
