+++ Mage.Sets/src/mage/cards/a/AyeshaTanaka.java
@@ -16,6 +16,7 @@ import mage.constants.SubType;
 import mage.constants.SuperType;
 import mage.constants.Zone;
 import mage.filter.FilterAbility;
+import mage.filter.FilterStackObject;
 import mage.filter.predicate.ability.ArtifactSourcePredicate;
 import mage.target.common.TargetActivatedAbility;
 
@@ -25,7 +26,7 @@ import mage.target.common.TargetActivatedAbility;
  */
 public final class AyeshaTanaka extends CardImpl {
 
-    private final static FilterAbility filter = new FilterAbility("activated ability from an artifact source");
+    private final static FilterStackObject filter = new FilterStackObject("activated ability from an artifact source");
 
     static {
         filter.add(new ArtifactSourcePredicate());
