+++ Mage.Sets/src/mage/cards/b/BrownOuphe.java
@@ -14,6 +14,7 @@ import mage.constants.CardType;
 import mage.constants.SubType;
 import mage.constants.Zone;
 import mage.filter.FilterAbility;
+import mage.filter.FilterStackObject;
 import mage.filter.predicate.ability.ArtifactSourcePredicate;
 import mage.target.common.TargetActivatedAbility;
 
@@ -23,7 +24,7 @@ import mage.target.common.TargetActivatedAbility;
  */
 public final class BrownOuphe extends CardImpl {
 
-    private final static FilterAbility filter = new FilterAbility("activated ability from an artifact source");
+    private final static FilterStackObject filter = new FilterStackObject("activated ability from an artifact source");
 
     static {
         filter.add(new ArtifactSourcePredicate());
